+++ integration_test/test/cloud_behavior_test.dart
@@ -20,6 +20,7 @@ import 'dart:convert';
 
 import 'package:functions_framework/serve.dart';
 import 'package:functions_framework/src/run.dart';
+import 'package:google_cloud/constants.dart' as cloud_constants;
 import 'package:google_cloud/google_cloud.dart';
 import 'package:hello_world_function_test/functions.dart';
 import 'package:test/test.dart';
@@ -74,7 +75,9 @@ void main() {
     count++;
     traceStart = 'trace_start$count';
 
-    headers = {cloudTraceContextHeader: '$traceStart/trace_end'};
+    headers = {
+      cloud_constants.cloudTraceContextHeader: '$traceStart/trace_end',
+    };
   }
 
   tearDown(() async {